Remove a node that has children if it's fully selected
This CL removes a node if it's fully selected, even though it has
children on deleting a selection.

While handling DeleteSelectionCommand, there is no chance to remove
a node itself when it has children, even though it's selected fully.
This change checks if the first position of `downstream_end_` anchor
node is after `upstream_start_` to know if it's included in the
selection. It removes the node, if it is selected from the first
position.

Bug: 331074432
Change-Id: I2c10904e8bc6b07cc6348449cf62b0cc71251f4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5404456
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1282356}
NOKEYCHECK=True
GitOrigin-RevId: 76b96062b866ed6a3741df9478a48c049b5906d2