[anchor] Support position-try-order, and the position-try shorthand
This CL implements position-try-order by retaining the IMCB on
OffsetInfo, collecting the successful candidates (and their
OffsetInfos) into a vector, and sorting that vector according to the
specified order.

This CL also effectively makes the position-try shorthand available
(status:experimental). This was only hidden previously because it
didn't make sense to support it without also supporting
position-try-order.

Note that whenever we have position-try-order:most-*, we'll now always
go through *all* the candidate styles (even if we find an early
candidate that fits).

Note also: when we're going through all candidates, we may now
populate `non_overflowing_scroll_ranges` with additional entries,
which could lead to more failures in AnchorPositionScrollData::
ValidateSnapshot. This is sub-optimal, but should be non-fatal.
We can optimize it later if necessary.

Bug: 332570740
Change-Id: I1c101ccffd5ef4dbce3a7128bfb0af6b72f3470c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5420296
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282454}
NOKEYCHECK=True
GitOrigin-RevId: bb849525348f5617dbb76949e7497ccd6d59b546