Fix and validate protobuf roll instructions
This updates the README.chromium to make it easier to roll protobuf, and
the result of trying to follow them are this changelist.

A few patches that were not generated in a previous submit are added to
patches/ and one is updated to apply cleanly. gen_chromium_file_lists.py
as well as BUILD.gn is updated to split up protoc into multiple source
sets for different languages. This avoids buildfile collisions between
different enum.cc files in proto_lib.

This should make the many future protobuf rolls slightly easier.

Bug: 328417294
Change-Id: Id61ea32489eeac69c5f8b2653ff08dcb198a423c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5420405
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282610}
NOKEYCHECK=True
GitOrigin-RevId: 141099563d180fc440ab84ee310e2c1561ba52d9