[Interop] Remove early return for complementary role
Remove the early return for complementary role in
IsDescendantOfLandmarkDisallowedElement(). The code after the early
return checks for a sectioning content element, which is required for
<aside>, according to [1].

The check for named <section> is moved so that kSection can be detected
in GetLandmarkIsNotAllowedAncestorRoles(). If the role is changed
earlier, that method would not work as intended.

Also, update AXObjectCacheImpl::ComputedRoleForNode(...) to return
the final role value. Previously, it returned an unfinalized role value
(not all rules had been applied yet), causing some web tests related to
region and section roles to fail.

[1] https://www.w3.org/TR/html-aam-1.0/#el-aside

Bug: 325975937, 41490744, 41493693
Change-Id: I9bad5be9b2d8d628a86ec3c739caeb522e370810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5341750
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282732}
NOKEYCHECK=True
GitOrigin-RevId: 2098980b7367231da23ca9a1101c18fc8daee8a6