Convert many callers of Length::IsPercentOrCalc().
By the time we have a Length, only calc() expressions involving
percentages that are resolved during layout, anchor(), or calc-size()
should still be represented as calculations.  So prior to anchor() and
calc-size(), IsPercentOrCalc() made sense conceptually.

This converts many callers to Length::HasPercent().  For some of these
callers the difference matters for calc-size() correctness;  for some it
does not currently matter (because the values can't contain calc-size())
and either one would be correct.  This conversion can also affect the
behavior when anchor() is used in a calc() expression (and percentages
are not).

Some other callers are converted to either (a) explicitly checking
Length::IsPercent() || Length::IsCalculated(), or (b) checking
!Length::IsFixed().

Callers that are using IsPercentOrCalc() to either:
 (a) optimize away computation of the percentage basis (max_value), or
 (b) optimize away handling of changes to the percentage basis
are left as-is.  These uses could be converted to a function that is
similar to HasPercent but also accounts for percentages like the one in
calc-size(any, 30%) that do depend on the percentage basis but do not
act like a percentage-typed value.  The function should probably be
renamed to account for this, and its implementation could perhaps be
changed as well, but this is not done in this CL.

Bug: 313072
Change-Id: I176abec7a6a03b3cf9975c5107ea32691f305d0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5410698
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282785}
NOKEYCHECK=True
GitOrigin-RevId: 420a92c3e960dc443b869120951c0ef864b9abd6