[css-anchor-position] Fix position fallback for scroll container of anchored
Scroll of the scroll container of an anchored element may affect
overflow status of the anchored element, though the scroll offset
doesn't contribute to the scroll adjustment. We need to invalidate
layout if such a scroll may affect the overflow status (which requires
NonOverflowingScrollRange to consider the scroll container), and
consider the scroll offset when calculating overflow of the anchored
element in TryCalculateOffset().

Bug: 331393773
Change-Id: I82f70b459f63d98a4b04899dee0c188bee2c8b46
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5416013
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282797}
NOKEYCHECK=True
GitOrigin-RevId: 868292a25e6f2eece006f4a4d05419c5bf4ef9d7