Spanify some MLBuffer code
https://crrev.com/c/5357633 introduced a UBSan regression:
https://ci.chromium.org/ui/labs/p/chromium/inv/build-8751722619949677969/test/ninja%3A%2F%2Fthird_party%2Fblink%2Frenderer%2Fcontroller%3Ablink_unittests%2FMLGraphTestMojo.WriteWebNNBufferTest%2FAll.WebNNService/variant/603704313003e25d?resultIndex=0

This is because of a truly obnoxious bug in the C standard. See
https://davidben.net/2024/01/15/empty-slices.html for details. But since
we want to move things to bounds-checked span methods anyway, I've
taken the time to spanify things. The changes are:

- Instead of memcpy, prefer span::copy_from. Once we fix the STL's
  interaction with bounded output iterators (see
  https://github.com/llvm/llvm-project/issues/78771), std::ranges::copy
  will work too, but span::copy_from is easier.

- Instead of std::equal, prefer an API that doesn't allow for mixing up
  begin/end iterators. std::ranges::equal works, but base::span has
  operator==, which is even easier.

To support these, I had to make make mojo_base::BigBuffer mode
span-friendly. It has data() and size(), but it's not iterable, so the
deduction guide for base::span didn't work, just the ctor. Leaving the
question of that inconsistency for later, I've added the iterators
because we may as well make it act like other buffer types.

Bug: 40248746, 40284755
Change-Id: Ie98173f19f06b99063fc4e7c14d4862da58c60e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5420741
Commit-Queue: David Benjamin <davidben@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282937}
NOKEYCHECK=True
GitOrigin-RevId: 8297ca6c302e695b272f7180fabd82a943af6666