Add support for logical binary operators and cast
This change implements support for the remaining binary logical ops,
doing so uncovered an issue in that CoreML returns bool tensors
for these operators. In order to return uint8 that WebNN expects, the
cast operator needed to be implemented to convert from bool to uint8.

The cast operator in CoreML only supports casting to uint8 in macOS 14+,
the min supported OS version for WebNN on macOS needed to be updated to
14 because of this.

Finally tests needed to be updated because CoreML has limited datatypes
supported as features, i.e. the tensors at the Edge of a model cannot
be uint8/int8/uint32 in CoreML. Tests convert the graphs outputs to
int32 to workaround this constraint.

Bug: 41481333
Change-Id: Ieae72e0e2ecabfd5519ea3cfc97b0a253ec54742
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5409661
Reviewed-by: Phillis Tang <phillis@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Sushanth Rajasankar <Sushraja@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1282964}
NOKEYCHECK=True
GitOrigin-RevId: 70499b63a39f06900efb73dba9963c2c354e1035