[line-clamp] Hide out of flow positioned boxes after the clamp point
This patch makes out of flow positioned boxes that appear in the
layout tree after the clamp point (that is, in tree order after the
end of the last shown line) hidden for paint. This happens depending
entirely on the layout tree order, regardless of where that box is
positioned.

This is achieved by storing a flag in the various `*OofPositionedNode`
classes to mark the node as hidden for paint. This flag is then
checked when creating the constraint space for the layout of the box.

Bug: 40336192
Change-Id: Ia42751b2fa7e3fee89648e26600035241bda6299
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331654
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283041}
NOKEYCHECK=True
GitOrigin-RevId: 4b672b7b9ebd432ab88549db5391a39404659837