Set the ending selection to the end of the list child on outdenting
This CL updates the ending selection to the end of the list child if
the selected node is in a list item on outdenting.

OutdentRegion() runs OutdentParagraph() for each paragraph. If the
paragraph is in a list, it uses InsertListCommand::UnlistifyParagraph()
that moves all children of HTMLLIElement at once after adjusting
`start` and `end` with FirstPositionInNode and LastPositionInNode of
HTMLLIElement. After finishing UnlistifyParagraph(), OutdentRegion()
should know where to start the next paragraph so that it runs
OutdentParagraph(), again.
There is an adjustment for `end_of_next_paragraph` with the ending
selection, but it could be the position already moved if the previous
paragraph is in a list item.
This change sets the ending selection to be synchronized with the
moved position if the paragraph is in a list item.

Bug: 331841851
Change-Id: Ib023a6764d62f9579a9b0a035ca2429a2d3479f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5404376
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283059}
NOKEYCHECK=True
GitOrigin-RevId: 259a44fde2f6024a1063bcc7f92829c8f45ccf86