Fix LCP for transparent color texts
Currently LCP pipeline does not skip texts styled by color:transparent.
This could be abused to obtain a faster LCP time by using content which is not visible to the user

This CL is to skip the LCP measure for transparent texts.

Bug: 41481679
Change-Id: Ia16fdb0ababc45d8b7d6dc3262c7878da32c748d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5398565
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283135}
NOKEYCHECK=True
GitOrigin-RevId: 0dfcd2054dc2191a7a4b1e4a71036fd6769706b8