Add use counter for animations with static properties.
Animations with static properties can be optimized out if all of the
properties are static, or run on the compositor if the only remaining
(dynamic) properties are supported on the compositor. The primary
source of static properties is shorthand expansion, where for example,
background is expanded to background-position background-color etc.
We cannot directly track when shorthand expansion has been applied in
an animation since the expansion occurs upstream of the animation code.

Bug: 332498723
Change-Id: I9c9b8f9746e52f6ad44f35eb7fdd9f6630bc7700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5426355
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283264}
NOKEYCHECK=True
GitOrigin-RevId: ab231cbe7fe00ff7328c08037c6ab385d1e9ea36