Add interest invokers "gain interest" concept and defaults
This patch adds handling for focus and hover to trigger the
"gain interest" concept.

Additionally implements the interest action to trigger a popover to
open.

This does not implement any delays based on CSS properties.

This also does not implement the "loseinterest" concept.

Bug: 326681249
Change-Id: I47e9ddfe5199b5a3ff2884e0bb3b4e19cdbdeaf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5340116
Commit-Queue: Luke <lwarlow@igalia.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283340}
NOKEYCHECK=True
GitOrigin-RevId: f12fa89204d091604222d1e8a9a1ecf9e1611baf