Add finalized_record_id to UMA log records.
This CL adds a(nother) client-managed id that functions as a sequence number for the log record. Clients are expected to maintain a counter, incrementing it (by 1) for each record they finalize, and populate this field. Each record_id is expected to be unique when scoped to a given client_id. It is further expected that the record_id values associated to a given client_id form an uninterrupted numeric sequence (i.e. k, k+1, k+2, k+3, ...), allowing for the detection of dropped, or otherwise lost, finalized records.

Bug: 331406952
Change-Id: I50f19bfaa0983f910bccbd6ff2b0064d6f7e2311
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5401477
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Roger McFarlane <rogerm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283351}
NOKEYCHECK=True
GitOrigin-RevId: c183b2e3954357346237e2f758edca08e9099504