[css-anchor-position] Don't use LayoutBox::FindTargetAnchor during layout
FindTargetAnchor() asks the container for PhysicalAnchorQuery which
is available only after layout of the container, so we should not use
it during layout of the anchored element.

A previous similar test
position-visibility-anchors-visible-both-position-fixed.tentative.html
is passing regardless of the issue because of the side effect of
crbug.com/333107112. The test added in this CL doesn't "benefit" from
that side effect.

Bug: 329703412
Change-Id: I24d1d17c17d14dde93878e79783e893762ee54d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5430556
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283534}
NOKEYCHECK=True
GitOrigin-RevId: 4ae06f3f531870b642dea8798cea8182c2867cb3