[IdentifiabilityStudy] Improve performance of canvas helper
This CL improves the performance of 2d canvas when the Identifiability
Study is enabled. The main idea is to bypass the
IdentifiabileTokenBuilder and instead compute the hash inside
IdentifiabilityStudyHelper directly. This allows to skip some checks.

In particular, since the canvas IdentifiabilityStudyHelper computes an
incremental hash of IdentifiableTokens (which is 8 bytes), and the
hash is computed every 64 bytes, there is no need to check for
alignment every time.

A unit test ensures that the computation gives the same hash as before.

Bug: 328278696
Change-Id: Ib1eb5ef810663f7a08d0deeb64729e83f0ae0394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5355738
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1283732}
NOKEYCHECK=True
GitOrigin-RevId: a9462e62f79cff697716c43ce94e504904dc99ad