IDB: fix flaky test cursor-request-cycle.html
These two nearly identical tests both relied on a particular ordering
between Oilpan GC and IDB event dispatch. In reality, the txn is free
to commit as soon as a new request is not issued in a previous request's
onsuccess, so this test is racy. It was failing approximately 2-3% of
the time on a local Linux build, and passing 100% after this fix.

It's not clear why this suddenly became flaky. Although the failures
more or less align with the addition of IndexedDBShardBackingStores
to the fieldtrial testing config, which adds the flag to bots' test
runs, the flakiness appears to be independent of the status of that
feature.

Fixed: 331291856
Change-Id: I36a2f09f8e27bbd0e9c522eaa4638d0a836da4c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5435605
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284189}
NOKEYCHECK=True
GitOrigin-RevId: 5604c3633b666a6e98328c889000be78da285704