Fix for flaky WPT crash due to unresponsive renderer
The way we were obtaining the world when generating the call stack
is causing some wpt tests that are already flaky due to "renderer
unresponsive" to crash. I believe this is because
`DOMWrapperWorld::Current(isolate);` goes down a code path where
`ScriptState::From` is called for a context that does not have valid
 embedder data in the embedder field.

`ExecutionContext::From(context);` avoids this by using `ScriptState::MaybeFrom(context);` which is why we're not seeing the crash on that line.

Bug: 331454368
Change-Id: I4fdd0601a7c9566c681a70c4a9c00f1e9a52bd2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402897
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Issack John <issackjohn@microsoft.com>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284194}
NOKEYCHECK=True
GitOrigin-RevId: 6f1b19fadb30abeb0c275863d2e793d5ecc3281a