RubyLB: TakeRubyColumnList() should always move ruby_column_list_
Before this CL, TakeRubyColumnList() returned a reference. So if we
didn't apply move operations to the returned reference,
`ruby_column_list_` was not moved.  It's confusing.

This CL updates TakeRubyColumnList() so that it returns a value instead
of a reference, and it always moves ruby_column_list_.

This CL has no production behavior changes, and is for code health.

Bug: 324111880
Change-Id: I2174d1cf3cc964d0b0b9e419625c2bb13d28b9d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5435100
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284257}
NOKEYCHECK=True
GitOrigin-RevId: 911e0726e8a4b81c4b21608f2e6b60125319110e