[PEPC]: Reset mojo receiver when PEPC is detached from layout tree
Given that the receiver has the same lifetime as PEPC, the JS garbage
collector will maintain the element for an arbitrary time even when the
element is detached from layout tree, consuming the quota of PEPC
instances for the page (currently limited to two per page). As a result,
new PEPCs will either be prevented from being constructed or their
construction will be delayed (until JS GC frees the previous PEPC
instances).

To address the above issue, we will bind the mojo if the element is
attached to the DOM. Conversely, if the element is detached from the
DOM, we will reset the mojo connection, separate from JavaScript's
garbage collection schedule.

Bug: 1462930
Change-Id: I14b6ef2c055d7f1e24f9ee78f4573cfd3745683a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5360897
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Thomas Nguyen <tungnh@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284367}
NOKEYCHECK=True
GitOrigin-RevId: f3541b24bd7457482c0684c837fbe8e01de13e35