[line-clamp] Hide floats after the clamp point
This patch makes floats that appear in the layout tree after the clamp
point (that is, in tree order after the end of the last shown line)
hidden for paint. This depends entirely on the layout tree order,
regardless of whether the float extends beyond the last line.

This is achieved by storing a flag in `UnpositionedFloat` to mark the
float as hidden for paint, which is then checked when creating the
constraint space for the float.

Bug: 40336192
Change-Id: I8216289461e9390ecb99454c2e81f889e5ca8489
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5425036
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284793}
NOKEYCHECK=True
GitOrigin-RevId: ee7debcb104b0224c0cc0c353aa55f856565eac3