Rework SVGResourceDocumentContent to be the ResourceClient
This separates the load-cycle of the underlying Resource from the load-
cycle of the SVGResourceDocumentContent. In a future CL they will no
longer be the same because additional steps need to be taken before
loading can be considered completed in some cases (c.f SVGImage and the
loading of data URL resources).

A new TextResource resource subclass is introduced -
SVGDocumentResource - to manage the loading and essentially do the work
of the ResourceClient. It feeds data and status updates into the
associated SVGResourceDocumentContent.

Bug: 40134477
Change-Id: Ide967d0d4a315c9205bdc86ece646c5726e8ab19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5419515
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285089}
NOKEYCHECK=True
GitOrigin-RevId: 4fd78d8e5ba3a306d8447b75b293acbad6e2ebfd