Implement -internal-appearance-auto-base-select()
This patch implements -internal-appearance-auto-base-select(), a CSS
function that can be used in UA stylesheets which takes two values. If
the computed style of the element is not appearance:base-selet, then the
first value is used. Otherwise, the second value is used.

This patch only uses it for background-color on <select>, but I will try
using it for additional properties like border next.

Bug: 1511354
Change-Id: I70fea56ef8a8c931418872039f1df38e5324f55f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5389566
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285298}
NOKEYCHECK=True
GitOrigin-RevId: cb322e133d218f042d6d7b14eec395803f80c026