[anchor] Disable AnchorResults and persistent try sets
Unfortunately this cache causes fundamental invalidation problems.
This CL disables the cache to fix the invalidation problem, but we
retain the code for now, while we investigate alternatives.

The cache is still used in a limited way to solve a problem related
to ::backdrop in the top layer, see documentation in
style_resolver_state.cc for more.

Bug: 333608683
Change-Id: Ia7277ea031a574ef794ceed6390d52fbade25dc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5442875
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285659}
NOKEYCHECK=True
GitOrigin-RevId: e4692f2b47dc3fdc9846ac40c27ddf8249e0ac7e