Speculative fix for AnchorElementMetricsSender::NavigationStart crash
Crash reports suggest we're running through the DCHECK. We switch the
uses of GetTopDocument in AnchorElementMetricsSender to use
GetSupplementable instead. In order for AnchorElementMetricsSender
to be called with a given anchor, its top document is necessarily the
same document associated with the AnchorElementMetricsSender.

Bug: 333319337
Change-Id: Ice4c95ea25374438a87b6aacdd247cfafc448521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5439603
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285887}
NOKEYCHECK=True
GitOrigin-RevId: 51abc6635f19218ee46b7780199620415b031faf