Disallow usage of canvas GPUTexture after endWebGPUAccess.
After endWebGPUAccess, the GPUTexture is conceptually owned by the
canvas--and eventually, it should _actually_ have its ownership
transferred to the canvas.

We can enforce that the GPUTexture is no longer accessible from
JavaScript by calling destroy() on it. Unfortunately, there is (by
design) no way to directly test that a GPUTexture has been
destroyed, so we test this indirectly instead. The test attempts
to read and write from the texture after endWebGPUAccess is called,
and expects GPUValidationErrors to occur.

Bug: 41490345
Change-Id: I6501e64c73e679a572eeb2a994ff51920539a9c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5439858
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Auto-Submit: John Stiles <johnstiles@google.com>
Cr-Commit-Position: refs/heads/main@{#1285935}
NOKEYCHECK=True
GitOrigin-RevId: 028f06b0b9494df0de196d46e9f4b5d8746ccc1b