Disable Component Update for Widevine on LaCros
As Component Update currently doesn't work for Widevine on LaCros, this
disables it. Only visible impact is that Widevine is no longer listed
on the chrome://components page. Widevine will still be used as it
bundled with LaCros.

Verified by checking that Widevine content plays and that it is no
longer listed on the chrome://components page.

Bug: b/332962687
Test: see above
Change-Id: Ibf0d34c7e1c32966a3389f9751712e62fcbddd4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445806
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285986}
NOKEYCHECK=True
GitOrigin-RevId: ca53cbde1728fbc0a30d1acb30357be55a18e26a