[anchor] Invalid anchor()/anchor-size() without fallback becomes IACVT
A recent CSSWG resolution changed the "missing fallback behavior"
from always returning 0px, to instead behaving as invalid at computed-
value time [1]. This is the first time a calc() result can become IACVT,
so we need to make some adjustments for this to be possible.

The approach in this CL does not have the best performance, but it gets
the job done without a massive calc()/ApplyValue code refactor.
In StyleCascade, at the place we currently resolve var() substitutions,
we also check if we need to "resolve" any CSSMathFunctionValues into
IACVT (unset). This is required when the CSSMathFunctionValue contains
anchor() or anchor-size() functions without fallbacks, and when
evaluating any of those produces no result. In order to figure this out,
we have to actually evaluate the queries. This unfortunately means that
we evaluate the queries twice for the time being: once for the validity
check, and again for the real calculation immediately after.

We had an existing function InvolvesAnchorQueries which traversed the
expression tree in full every time, but I changed this to instead
return a boolean stored on the CSSMathExpressionNode itself. Since the
approach is identical to HasComparisons, I renamed to HasAnchorFunctions
for consistency. The reason for this change is to avoid an expensive
InvolvesAnchorQueries check for every CSSMathFunctionValue seen
in the cascade.

Evaluating calc() values with anchor*() inside, even to just check for
validity, needs to happen on scoped values (CSSValue::IsScopedValue).
Previously, we'd scope-ify the value immediately before ApplyPhysical-
Property (which ultimately evaluates any anchor*() functions), but now
we can risk evaluating anchor* () sooner than that. Therefore I had to
split out the tree-scope-ensuring code, and use that during the Resolve
step for the CSSMathFunctionValue.

Finally, the static StyleCascade::Resolve function, which is an API for
resolving arbitrary CSSValues against the computed values of some
element, used kAuthor (by arbitrary choice) as the value's origin.
This conflicts with EnsureScopedValue's reasonable assumption that any
cascade entry with kAuthor is backed by an actual MatchResult entry,
hence I changed this to kNone. This part of the CL should be a "no
behavior change" vs the existing code.

[1] https://github.com/w3c/csswg-drafts/issues/10005

Bug: 333720826
Change-Id: I7e02beb44981fd47a30981ef808af3ec18c1950a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5447517
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286010}
NOKEYCHECK=True
GitOrigin-RevId: 2db1fc3b7269dc1dd57dcb654a4dc5ea3a99a0a1