Use preloading ML model to enact moderate candidates
We add a new predictor type for the model and when the model produces
a result that is sufficiently confident, it enacts prerender or
prefetch candidates.

At this point, we have a number of feature parameters for whether
the model is actually active or just recording experimental prediction
data, and for the confidence thresholds. The thresholds have been
chosen arbitrarily for now.

When the model enacts candidates, we have it supersede the hover
heuristic. Otherwise, if the model decides not to enact, hovering
would naively enact it anyway. This is done on a per-page basis,
since even if the model is loaded, it may not run for some pages
(e.g. insecure http), so we continue to use the hover heuristic in
such cases.

Bug: 40278151
Change-Id: Iadb5412b672d3fded6c9f573025a31f1b57c7234
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5358129
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Auto-Submit: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Max Curran <curranmax@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286051}
NOKEYCHECK=True
GitOrigin-RevId: 842eb0a578d0e072d28d19fdd2bfd89e27f21528