RubyLB: Introduce kRubyLinePlaceholder InlineItem
It's added to the beginning and the end of ruby lines.
Inline positioning takes into account of kRubyLinePlaceholder items,
but FragmentItems are not created for them.

This is a preparation to implement the ruby-align:space-around behavior.

Bug: 324111880
Change-Id: I210dab20b04b392a4281de651332e58fe3ca9f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5446113
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286300}
NOKEYCHECK=True
GitOrigin-RevId: 713c2f855cb002779dfb9250c290f632c5c2cc7d