[anchor] Mark PositionedMovementLayout for null-AnchorEvaluator styles
Regular style recalcs (that are not interleaved from OOF layout)
evaluate anchor*() functions without an AnchorEvaluator, which means
the fallback behavior is always taken. This means if we initially have
some style with left:10px, and then (in a separate lifecycle) change
the specified value to left:anchor(--a right,10px), the ComputedStyle
diffing will not be able to tell that something changed (because it
didn't: anchor() resolves computed-value time). Yet we still need
to invalidate layout in this situation, otherwise we'll never enter
the interleaving OOF recalc and evaluate the anchor() function
for real.

Therefore, this CL (effectively) marks for PositionedMovementLayout
in this situation. I do not actually set the the flag in the diff,
however, because the decision to invalidate is based on the new
style alone (i.e. no actual diffing).

Fixed: 333858786
Change-Id: I49d08c0551ae2ce54d08b6ec947e56c5514c0a61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445690
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286396}
NOKEYCHECK=True
GitOrigin-RevId: cf808620adbe4bd74d0bb2d3173610eb280ec756