Fix popover-focus-2.html WPT
We started failing this WPT after this PR was merged:
https://github.com/web-platform-tests/wpt/pull/43146

The assert we are failing is about which element is focused after
disabling a button, and we are not focusing that disabled button, so I
tweaked the assertion to say that we should not be focusing the button.

I'd be worried that we are diverging from firefox behavior here, but
based on wpt.fyi firefox is already failing the test anyway.

Bug: 325981758
Change-Id: I66d2d86e11546ce4ca11db001b85e5f36d5bccde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5403252
Auto-Submit: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286654}
NOKEYCHECK=True
GitOrigin-RevId: 1fb84c18466468e0292863059d21bd24606dcaae