[anchor-pos] Fix the non-auto-inset calculation for position fallback.
Previously we didn't consider inset-area correctly (and assumed that
everything was non-auto).

To fix this delay the coercing of InsetAreaOffsets until later, then
use that information to determine if its "auto" or not.

This doesn't fix the general position-fallback issue, just a pre-step

Bug: 333614330
Change-Id: I0fee1bdbaf1c28cc9461db641cbb78dc5009db05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5450176
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286787}
NOKEYCHECK=True
GitOrigin-RevId: acc913a3610605a42886edcf5cea2f746c2aff72