[anchor] Implement CSSPositionTryDescriptors
CSSPositionTryDescriptors is a CSSStyleDeclaration which restricts
which properties are valid, which is a new concept in Blink.
This CL implements this by:

 - Adding a new parsing mode which is stored on the CSSPropertyValueSet
   within the rule. This ensures that we can't smuggle in invalid
   properties using cssText.
 - Filtering properties during setProperty. This ensures that e.g.
   setProperty('color', ...) doesn't work.

Somehow, CSSPositionTryDescriptor is not inheriting the anonymous
named getter/setter from the parent, which is actually what we want
here. (If this is a bug or something and is changed later, the WPT will
fail).

The indexed property getter also doesn't seem to work. This is not
covered by the tests currently. The problem is tracked by
Issue 333965945.

Change-Id: I5fa1b600ed50633a9eb9aa8cfeb43726d64ecf90
Bug: 333967631, 333965945
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5447110
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1287213}
NOKEYCHECK=True
GitOrigin-RevId: 10729d008a3cc659f210e226d7d3110d2ae0c40f