[CompositeClipPathAnimation] Refactor test only branches
This change follows up on earlier work which enhanced testing for composited clip path animations by removing the dependence on the need for a mocked class.

The test only branches that test for the presence of the generator have been moved into the paint worklet proxy client, making the generator function both within and without the test environment.

A change reflecting this for composited bgcolor animations will be added as a followup to this change.

Bug: 325517328

Change-Id: I9803a689a5c9359e4e91bfe7d9a671186ac6bdd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5362607
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1287477}
NOKEYCHECK=True
GitOrigin-RevId: 6f50bc159201186145ffe46847df0afa7bfb255e