Fix popover-light-dismiss.html
We are failing the last subtest in popover-light-dismiss which clicks
and drags from the main frame to an iframe because chromedriver is
telling chrome that we are doing pointerUp in both the button in the
main frame and the body of the iframe at the same time. I'm not sure
why, but combining the actions sequences and using the main frame's
document as the origin for the pointerUp fixes this.

The failure only occurs with run_wpt_tests.py, not run_web_tests.py.

Since writing the above message, another test case was appended which
doesn't include an iframe, and the same fix of using one action sequence
was needed there too, so there must be some chromedriver bug which
doesn't even involve iframes. I verified with internal debug logging
that the new action sequences do in fact exercise the code which needs
to be tested, at least in chromium.

Bug: 325981758
Change-Id: I992a01a6baf78a7701da3dd8df76788c93ed0aaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5402096
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1287670}
NOKEYCHECK=True
GitOrigin-RevId: 02d797ea20afb35ef326458b54de758da8f492a5